requirements.txt
beautifulsoup4
requests
lxml
langchain
langchain-text-splitters
langchain-mistralai
langchain-core
sentence-transformers
qdrant-client
numpy
pydantic
python-dotenv
